ThinkPython/Chapter_19/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
canvasvg==1.0.5
certifi==2023.5.7
charset-normalizer==3.1.0
googlesearch-python==1.2.3
idna==3.4
Pillow==9.5.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
swampy==3.0.1
tkhtmlview==0.2.0
tzdata==2023.3
urllib3==2.0.3
